+++ cfg4py/__init__.py
@@ -14,7 +14,7 @@ from cfg4py.core import (
 
 __author__ = """Aaron Yang"""
 __email__ = "code@jieyu.ai"
-__version__ = "0.9.2"
+__version__ = "0.9.3"
 
 
 def get_instance():